# optimize NormHead for Baichuan2