## Description


### 1. Why the change?
To optimize the memory footprint and latency for Baichuan2

### 2. User API changes
None

### 3. Summary of the change 

At inference time, `NormHead().forward` does weights normalization once (i.e. first time enters forward) and then act just like a normal `Linear` layer. So before applying low-bit optimizations, we 1) do weight normalization and 2) replace it with a standard `Linear` layer. Then it can be handled just like normal `Linear` layer.  

#### Model architecture (simplified):
Before: 
```python
BaichuanForCausalLM(
  (model): BaichuanModel(
      ...
  )
  (lm_head): NormHead()
)
```
After: e.g. 
7B
```python
BaichuanForCausalLM(
  (model): BaichuanModel(
      ...
  )
 (lm_head): LowBitLinear(in_features=4096, out_features=125696, bias=False)
)
``` 
13B
```python
BaichuanForCausalLM(
  (model): BaichuanModel(
      ...
  )
  (lm_head): LowBitLinear(in_features=5120, out_features=125696, bias=False)
)
``` 
### 4. Accuracy, Performance and Memory footprint Evaluation on CPU
Results collected on CPX-2 and it's just quick rough results from running the [Baichuan2 example](https://github.com/intel-analytics/BigDL/tree/main/python/llm/example/CPU/HF-Transformers-AutoModels/Model/baichuan2) 

#### Performance and Memory Footprint Improvement
lm_head nbytes=~1G bytes (fp32) for 7B model and lm_head nbytes=~2.5G(fp32) in 13B model as the hidden size is different. 
The memory footprint improvement is within expectation
 
|model|Memory Before|Memory After|Inference time Before|infernce time After|
|---|---|--|---|---|
|7B|~7G|~5.8G|12.8 s|5.9s| 
|13b|~14G|~11.8G|17.6 s|13.9s|


#### Accuracy seems not affected

|7B original|7B After optimized|
|--|--|
|<human>AI是什么？ <bot>AI是人工智能（Artificial Intelligence）的缩写，它是指让计算机或其他设备模拟和执行人类智能的技术。AI可以应用于各种任务，如|<human>AI是什么？ <bot>AI 是人工智能（Artificial Intelligence）的缩写，它是指让计算机或机器模拟和执行人类智能的技术。AI 系统可以通过学习、|

|13B original|13B After optimized|
|--|--|
|<human>AI是什么？ <bot>人工智能（AI）是指由计算机系统或其他数字设备模拟、扩展和增强人类智能的科学和技术。它涉及到多个领域，如机器学习、计算机视觉、|<human>AI是什么？ <bot>人工智能（Artificial Intelligence）是指由计算机系统或其他数字设备模拟、扩展和增强人类智能的技术。它涉及到多个领域，如机器学习、计算机视觉|

### Performance and Memory footprint on GPU (TODO)