@shane-huang 2023-10-18
Merge for now. Will collect comprehensive cpu/gpu data and continue optimize if necessary.